+++ src/leetcode/medium/No1922.java
@@ -10,7 +10,7 @@ public class No1922 {
             int[] primeNumber = new int[]{2, 3, 5, 7};
             int[] evenNumber = new int[]{0, 2, 4, 6, 8};
             int ans = 0;
-            return (int) (modPow(20, n / 2) * (n % 2 == 1 ? 5 : 1) % K_MOD);
+            return modPow(20, n / 2) * (n % 2 == 1 ? 5 : 1) % K_MOD;
         }
 
         private int modPow(int x, long n) {
